aliases tests refactored for snowflake support
