+++ phpunit.xml.dist
@@ -17,16 +17,20 @@
 
 	<testsuite name="backend-mysql">
 		<directory>tests/Backend/Common</directory>
+		<file>tests/Backend/Aliases/SimpleAliasTest.php</file>
 		<directory>tests/Backend/Mysql</directory>
 	</testsuite>
 
 	<testsuite name="backend-redshift">
 		<directory>tests/Backend/Common</directory>
 		<directory>tests/Backend/Redshift</directory>
+		<file>tests/Backend/Aliases/CustomSqlAliasTest.php</file>
 	</testsuite>
 
 	<testsuite name="backend-snowflake">
 		<directory>tests/Backend/Common</directory>
+		<file>tests/Backend/Aliases/SimpleAliasTest.php</file>
+		<file>tests/Backend/Aliases/CustomSqlAliasTest.php</file>
 	</testsuite>
 
 </phpunit>
